# site/public/openapi.yaml
openapi: 3.1.0
info:
  title: Cadastrum Public API
  version: 1.2.0
  description: |
    Cadastrum API — gayrimenkul fiyat tahmin, emsal sorgu ve risk verisi.
    Kurumsal Pro tier'a özel programmatic erişim.

    Token formatı: `cdrm_<32-hex>`. /v1/api/tokens endpoint'inden üretilir.
    Rate limit varsayılan: 60 req/dakika/token (oluştururken özelleştirilebilir).

    **Değişiklik geçmişi:**
    - v1.2.0: /fiyat/trend, /fiyat/toplu-ozet, /fiyat/toplu-ilce-ozet, /sorgu endpoint'leri eklendi.
              /fiyat/{il,ilce} endpoint şemaları tamamlandı. Seed endpoint'leri (internal) belgelendi.
    - v1.1.0: İlk public release.
  contact:
    name: Cadastrum
    email: eparlak996@gmail.com
servers:
  - url: https://cadastrum-api.cadastrum-tr.workers.dev
    description: Production (Cloudflare Worker)
  - url: https://api.cadastrum.com.tr
    description: Production (custom domain)
security:
  - ApiKeyAuth: []
paths:
...
  /v1/auth/kayit:
    post:
      tags: [Auth]
      summary: Yeni kullanıcı kaydı
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, sifre]
              properties:
                email: { type: string, format: email }
                sifre: { type: string, minLength: 8 }
                ad: { type: string }
      responses:
        "200":
          description: Kayıt başarılı
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthCevap' }
        "409":
          description: Email zaten kayıtlı
  /v1/auth/giris:
    post:
      tags: [Auth]
      summary: Giriş yap + JWT al
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, sifre]
              properties:
                email: { type: string, format: email }
                sifre: { type: string }
      responses:
        "200":
          description: Giriş başarılı
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthCevap' }
        "401":
          description: Hatalı email/şifre
        "429":
          description: Rate limit (10 deneme/dakika)

# ── Components ────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Kurumsal Pro tier API anahtarı.
        Format: `cdrm_<32-hex>`. /v1/api/tokens endpoint'inden oluşturulur.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token — /v1/auth/giris endpoint'inden alınır.
        /v1/api/tokens CRUD işlemleri için gereklidir.
  schemas:
...
    AuthCevap:
      type: object
      properties:
        token: { type: string, description: JWT (30 gün geçerli) }
        kullanici:
          type: object
          properties:
            id: { type: integer }
            email: { type: string }
            ad: { type: string, nullable: true }
            tier: { type: string, enum: [free, pro, pro_plus, kurumsal] }
            tierBitis: { type: integer, nullable: true }
            emailDogrulandi: { type: boolean }
...